add extra probe arameters to emulator rework probe configuration handling in the mobile app normalize parameters input as float/short change emulator data file

+++ android-app/app/src/main/java/com/echopen/asso/echopen/model/Data/DeviceConfiguration.java
@@ -5,17 +5,17 @@ package com.echopen.asso.echopen.model.Data;
  */
 public class DeviceConfiguration {
 
-    private double mR0; // measure initial depth in m
-    private double mRf; // measure final depth in m
+    private float mR0; // measure initial depth in m
+    private float mRf; // measure final depth in m
     private float mProbeSectorAngle; // measure sector angle in degree
-    private double mSamplingFrequency; // sampling frequency in Hz
+    private float mSamplingFrequency; // sampling frequency in Hz
     private int mNbLinesPerImage;
     private int mNbSamplesPerLine;
     private String mProbeCinematicName;
     private float mEchoDelay; // echo delay measurement in milliseconds
     private float mDecimation; // echo decimation
 
-    public DeviceConfiguration(double iR0, double iRf, float iProbeSectorAngle, double iSamplingFrequency, int iNbLinePerImage, int iNbSamplesPerLine, String iProbeCinematicName, float iEchoDelay, float iDecimation) {
+    public DeviceConfiguration(float iR0, float iRf, float iProbeSectorAngle, float iSamplingFrequency, int iNbLinePerImage, int iNbSamplesPerLine, String iProbeCinematicName, float iEchoDelay, float iDecimation) {
         mR0 = iR0;
         mRf = iRf;
         mProbeSectorAngle = iProbeSectorAngle;
@@ -27,11 +27,11 @@ public class DeviceConfiguration {
         mDecimation = iDecimation;
     }
 
-    public double getR0(){
+    public float getR0(){
         return mR0;
     }
 
-    public double getRf(){
+    public float getRf(){
         return mRf;
     }
 
@@ -39,7 +39,7 @@ public class DeviceConfiguration {
         return mProbeSectorAngle;
     }
 
-    public double getSamplingFrequency(){
+    public float getSamplingFrequency(){
         return mSamplingFrequency;
     }
